Implement HTTP and TCP monitoring checks
- Added `check_http_target` function for performing HTTP checks, including SSL certificate validation and response assertions.
- Introduced `check_tcp_port` function for TCP connectivity checks.
- Created a new module `checks` to organize various monitoring checks.
- Updated the main monitoring loop to include HTTP and TCP checks, handling results and logging appropriately.
- Enhanced `TargetStatus` struct to store detailed check results, including response times and certificate validity.
- Implemented historical check result tracking for uptime and response time metrics.

+++ src/api.rs
@@ -148,7 +148,7 @@ async fn badge_handler(
                 message,
                 color,
                 response_time,
-                Some(status.uptime_percentage_24h),
+                Some(status.uptime_percentage_24h as f32),
             );
             
             HttpResponse::Ok()
